🚚 Restructured files of the CLI
