+++ cli/src/main.rs
@@ -1,23 +1,17 @@
 use clap::Parser;
 use crate::arguments::Arguments;
-use crate::arguments::cli_operation::CliOperation;
-use crate::operations::{manage_user, start};
+use crate::arguments::cli_commands::CliCommands;
+use crate::commands::{run_command};
 
 pub(crate) mod error;
 pub(crate) mod arguments;
-pub(crate) mod operations;
+pub(crate) mod commands;
 pub(crate) mod prelude;
-pub(crate) mod init;
 
 #[tokio::main]
 async fn main() {
     let arguments = Arguments::parse();
 
-    match arguments.operation {
-        CliOperation::Start => {
-            start().await;
-        }
-        CliOperation::User(options) => manage_user(options).await
-            .unwrap()
-    }
+    run_command(arguments.command)
+        .await;
 }
